packages/sage-benchmark/src/sage/benchmark/benchmark_anns/requirements.txt
# Requirements for benchmark_anns

# Core dependencies
numpy>=1.20.0
pyyaml>=5.4.0
pandas>=1.3.0

# Optional: PyCANDYAlgo for Faiss bindings
# Note: This may need to be installed separately or built from source
# PyCANDYAlgo

# For development
pytest>=6.0.0
